requirements.txt
# Core Flask stack
Flask==3.0.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
Flask-Login==0.6.3
Flask-SocketIO==5.5.1
Flask-SSE==1.0.0

# SQL drivers
mysqlclient==2.2.7
PyMySQL==1.1.1
psycopg2==2.9.10
psycopg2-binary==2.9.10

# ORM & Migrations
SQLAlchemy==2.0.38
alembic==1.15.2
Mako==1.3.10

# Security
cryptography==44.0.1
cffi==1.17.1
pycparser==2.22
PyJWT==2.10.1
itsdangerous==2.2.0

# Utility & Web
Werkzeug==3.1.3
Jinja2==3.1.4
MarkupSafe==3.0.2
click==8.1.7
typing_extensions==4.12.2
python-dotenv==1.1.0

# Email & Validation
dnspython==2.7.0
better-profanity==0.7.0
dkimpy==1.1.8

# Deployment
gunicorn==21.2.0
greenlet==3.1.1

# Optional
blinker==1.9.0
cachelib==0.9.0
